requirements.txt
Django~=5.1.3
drf-spectacular~=0.27.2
fastapi~=0.115.5
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2
django-environ==0.11.2
django-extensions==3.2.3
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
drf-spectacular-sidecar==2024.7.1
django-filter==24.3
pyotp==2.9.0
requests==2.32.3
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.2.0